fix(Plant): Change the response and http messages when a user tries to delete an image of plant or create a plant

+++ src/controllers/plant.controller.ts
@@ -1,31 +1,39 @@
-import { NextFunction, Request, Response } from "express";
-import {findLimitedPlantsByUserIdAndSpeciesId,findOnePlant,createPlant,addImageWithPlantId,deleteImageWithPlantId ,deletePlantsWithPlantsId,updatePlantWithPlantId} from "../queries/plant.queries"
-import { findOneSpecies, } from "../queries/species.queries";
-import  mongoose from 'mongoose';
-import { newImage } from "./image.controller";
-import { deleteImage ,getImageById} from "../queries/image.queries";
+import {NextFunction, Request, Response} from "express";
+import {
+    findLimitedPlantsByUserIdAndSpeciesId,
+    findOnePlant,
+    createPlant,
+    addImageWithPlantId,
+    deleteImageWithPlantId,
+    deletePlantsWithPlantsId,
+    updatePlantWithPlantId, getOnePlantById
+} from "../queries/plant.queries"
+import {findOneSpecies,} from "../queries/species.queries";
+import mongoose from 'mongoose';
+import {newImage} from "./image.controller";
+import {deleteImage, getImageById} from "../queries/image.queries";
 import {
     getMyPlantsValidation,
-    getOneOfMyPlantValidation,
+    getOneOfMyPlantValidation, newPlantValidation,
     plantValidation
 } from "../database/validation/plant.validation"
 import * as fs from 'fs';
 import {API_HOSTNAME, API_VERSION, return400or500Errors} from "../utils";
 
-const limit:number = 5
+const limit: number = 5
 
 export const getMyPlants = async (req: Request, res: Response, _: NextFunction) => {
     try {
         const speciesId = (req.query.speciesId) ? String(req.query.speciesId) : ""
         const search = (req.query.search) ? String(req.query.search) : ""
         const page = (req.query.page) ? Number(req.query.page) : Number(1)
-        const order = (req.query.order == "ASC") ? 1  : -1
+        const order = (req.query.order == "ASC") ? 1 : -1
 
         await getMyPlantsValidation.validateAsync(req.query, {abortEarly: false});
 
-        const skip:number =  limit * page - limit;
-        const plants = await findLimitedPlantsByUserIdAndSpeciesId(req.user._id,speciesId,limit,skip,order,search)
-        if(plants){
+        const skip: number = limit * page - limit;
+        const plants = await findLimitedPlantsByUserIdAndSpeciesId(req.user._id, speciesId, limit, skip, order, search)
+        if (plants) {
             const result: any[] = []
             plants.forEach(item => {
                 result.push({
@@ -37,15 +45,15 @@ export const getMyPlants = async (req: Request, res: Response, _: NextFunction) 
             })
 
             res.status(200).json({result});
-        }else{
+        } else {
             res.status(404).json({
                 "field": ["error"],
                 "message": ["Not species found."]
             })
         }
-      } catch (e) {
+    } catch (e) {
         return400or500Errors(e, res)
-      }
+    }
 };
 
 
@@ -54,7 +62,7 @@ export const getOneOfMyPlant = async (req: Request, res: Response, _: NextFuncti
         await getOneOfMyPlantValidation.validateAsync(req.params, {abortEarly: false});
 
         const plant = await findOnePlant(req.params.plantId)
-        if(plant){
+        if (plant) {
             res.status(200).send({
                 _id: plant._id,
                 name: plant.name,
@@ -63,9 +71,9 @@ export const getOneOfMyPlant = async (req: Request, res: Response, _: NextFuncti
                 speciesInfo: API_HOSTNAME + "/api" + API_VERSION + "/species/" + plant.species._id
             })
         }
-      } catch (e) {
+    } catch (e) {
         return400or500Errors(e, res)
-      }
+    }
 };
 
 export const newPlant = async (req: Request, res: Response, _: NextFunction) => {
@@ -74,51 +82,51 @@ export const newPlant = async (req: Request, res: Response, _: NextFunction) => 
         const userId = req.user._id
         const speciesId = req.body.speciesId
 
-        await plantValidation.validateAsync(req.body, {abortEarly: false});
+        await newPlantValidation.validateAsync(req.body, {abortEarly: false});
 
         const species = await findOneSpecies(speciesId)
-        if (species){
-            const newPlant = await createPlant(speciesId,userId,plantName)
-            
+        if (species) {
+            const newPlant = await createPlant(speciesId, userId, plantName)
+
             res.status(200).send({
                 status: "success",
                 plantInfo: API_HOSTNAME + "/api" + API_VERSION + "/plant/" + newPlant._id
             });
-        }else{
+        } else {
             res.status(404).send({
                 field: ["error"],
                 message: ["Species not Found"]
             });
         }
-      } catch (e) {
+    } catch (e) {
         return400or500Errors(e, res)
-      }
-      
+    }
+
 };
 
 export const updatePlant = async (req: Request, res: Response, _: NextFunction) => {
-  try {
+    try {
 
-    await plantValidation.validateAsync(req.body, { abortEarly: false });
+        await plantValidation.validateAsync(req.body, {abortEarly: false});
 
-    const species = await findOneSpecies(req.body.speciesId)
-    const plant = await updatePlantWithPlantId(req.body);
-    if(species && plant){
-        res.status(200).json( {
-            status: "success",
-            plantInfo: API_HOSTNAME + "/api" + API_VERSION + "/plant/" + plant._id
-        });
-    }else{
-        res.status(404).send({
-            field: ["error"],
-            message: ["Species not Found"]
-        })
+        const species = await findOneSpecies(req.body.speciesId)
+        const plant = await updatePlantWithPlantId(req.body);
+        if (species && plant) {
+            res.status(200).json({
+                status: "success",
+                plantInfo: API_HOSTNAME + "/api" + API_VERSION + "/plant/" + plant._id
+            });
+        } else {
+            res.status(404).send({
+                field: ["error"],
+                message: ["Species not Found"]
+            })
+        }
+
+    } catch (e) {
+        return400or500Errors(e, res)
     }
 
-  } catch (e) {
-    return400or500Errors(e, res)
-  }
-    
 };
 
 export const addImageFromPlant = async (req: Request, res: Response, _: NextFunction) => {
@@ -135,7 +143,7 @@ export const addImageFromPlant = async (req: Request, res: Response, _: NextFunc
                     "status": "success",
                     "plantInfo": API_HOSTNAME + "/api" + API_VERSION + "/plant/" + newImagePlant._id
                 });
-            }else{
+            } else {
                 res.status(400).send({
                     field: ["error"],
                     message: ["PlantId is not Found."]
@@ -152,26 +160,29 @@ export const addImageFromPlant = async (req: Request, res: Response, _: NextFunc
     }
 
 };
-export const removeImageFromPlant= async (req: Request, res: Response, _: NextFunction) => {
+export const removeImageFromPlant = async (req: Request, res: Response, _: NextFunction) => {
     try {
-     
-        const plantId = req.params.plantId
-        const imageId = req.params.imageId
-        const image = await  getImageById(imageId)
-        if(image){
-        const newPlant = await deleteImageWithPlantId(image._id,new  mongoose.Types.ObjectId(plantId.trim()))
-        fs.unlinkSync("public/image/" + image.path);
-        await deleteImage(imageId)
-        res.status(200).send(newPlant);
-        }else{
-          res.status(404).send("Cet image n'existe pas");
+        const plant = await getOnePlantById(req.params.plantId)
+        const image = await getImageById(req.params.imageId)
+        if (image && plant) {
+            await deleteImageWithPlantId(image._id, req.body.imageId)
+            fs.unlinkSync("public/image/" + image.path);
+            await deleteImage(req.body.imageId)
+            res.status(200).send({
+                type: "success",
+                message: "File deleted"
+            });
+        } else {
+            res.status(404).send({
+                field: ["error"],
+                message: ["Plant or image not found"]
+            });
         }
-        
-  
-      } catch (e) {
-        res.status(404).send("error");
-      }
-      
+
+    } catch (e) {
+        return400or500Errors(e, res)
+    }
+
 };
 export const removePlant = async (req: Request, res: Response, _: NextFunction) => {
     try {
